implement options in gallery activity and fix problem with thumbnails resolution

+++ src/pl/pwr/sztuczneoko/ui/GalleryActivity.java
@@ -21,14 +21,18 @@ import android.app.AlertDialog;
 import android.app.Dialog;
 import android.app.DialogFragment;
 import android.app.ProgressDialog;
+import android.content.Context;
 import android.content.DialogInterface;
+import android.content.Intent;
 import android.content.pm.ActivityInfo;
 import android.content.res.Resources;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
+import android.net.Uri;
 import android.os.AsyncTask;
 import android.os.Bundle;
 import android.os.Environment;
+import android.util.DisplayMetrics;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.Menu;
@@ -37,7 +41,9 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.AdapterView;
 import android.widget.AdapterView.OnItemClickListener;
+import android.widget.EditText;
 import android.widget.GridView;
+import android.widget.LinearLayout;
 import android.widget.Toast;
 import android.widget.ViewSwitcher;
 import android.os.Build;
@@ -55,15 +61,23 @@ public class GalleryActivity extends soActivity {
     private MenuItem mItemRename;
     private MenuItem mItemSwitchDirectory;
     private String directory;
-    
+    private int width;
+    private int height;
+	private MenuItem mItemShowImg;
+	private MenuItem mItemFilterOptions;
+	private MenuItem mItemBTOptions;
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {		
 		super.onCreate(savedInstanceState);
+		DisplayMetrics displaymetrics = new DisplayMetrics();
+		getWindowManager().getDefaultDisplay().getMetrics(displaymetrics);
+		height = displaymetrics.heightPixels;
+		width = displaymetrics.widthPixels;
 		directory = Environment.getExternalStorageDirectory().getAbsolutePath()+"/soAppDir/myImages/";
 		setContentView(R.layout.activity_gallery);
 		setRequestedOrientation(ActivityInfo.SCREEN_ORIENTATION_PORTRAIT);
 		gridView = (GridView) findViewById(R.id.gridView);		
-        gridAdapter = new GalleryGridViewAdapter(this, R.layout.row_grid, imageItems,directory);
+        gridAdapter = new GalleryGridViewAdapter(this, R.layout.row_grid, imageItems,directory,height,width);
         gridView.setAdapter(gridAdapter);       
     }
 	@Override
@@ -72,7 +86,7 @@ public class GalleryActivity extends soActivity {
         OpenCVLoader.initAsync(OpenCVLoader.OPENCV_VERSION_2_4_3, this, mLoaderCallback);
 	}
 	public void sendButtonClick(View view){		
-		core.sendPhoto(this);
+		core.sendPhoto(this,directory);
 	}
 	public void setSelectedImg(ImageItem selectedImg,View v,int index) {
 		if(selectedImg.getImage()== null)return;
@@ -110,28 +124,73 @@ public class GalleryActivity extends soActivity {
     	mItemDelete = menu.add("usuń");
     	mItemRename = menu.add("zmien nazwe");
     	mItemSwitchDirectory = menu.add("zmien folder");
+    	mItemShowImg = menu.add("pokaz zdjecie");
+    	mItemBTOptions = menu.add("ustawienia bluetooth");
+    	mItemFilterOptions = menu.add("ustawienia filtracji");
         return true;
     }
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
-        if(selectedImg!=null){
-	    	if (item == mItemDelete) {
-	    		File file = new File(directory+selectedImg.getTitle());
-	    		file.delete();
-	    		gridAdapter.notifyDataSetChanged();
-	        } else if (item == mItemRename) {
-	        } else if (item == mItemSwitchDirectory) {
-	        	if(directory.matches(".*myImages.*"))
-	        		directory = Environment.getExternalStorageDirectory().getAbsolutePath()+"/soAppDir/myFilterImages/";
-	        	else
-	        		directory = Environment.getExternalStorageDirectory().getAbsolutePath()+"/soAppDir/myImages/";
-	        	gridAdapter = new GalleryGridViewAdapter(this, R.layout.row_grid, imageItems,directory);
-	            gridView.setAdapter(gridAdapter);   
-	        	gridAdapter.notifyDataSetChanged();
-	        }
-        }
-        return true;
+    	if (item == mItemDelete&&selectedImg!=null) {
+    		File file = new File(directory+selectedImg.getTitle());
+    		Log.d("del","delete file: "+directory+selectedImg.getTitle());
+    		file.delete();
+    		gridAdapter.notifyDataSetChanged();
+        } else if (item == mItemRename&&selectedImg!=null) {
+        	AlertDialog.Builder alertDialog = new AlertDialog.Builder(GalleryActivity.this);
+
+            alertDialog.setTitle("Zmien nazwe");
+
+            alertDialog.setMessage(selectedImg.getTitle());
+            final EditText input = new EditText(GalleryActivity.this);
+            LinearLayout.LayoutParams lp = new LinearLayout.LayoutParams(
+                                    LinearLayout.LayoutParams.MATCH_PARENT,
+                                    LinearLayout.LayoutParams.MATCH_PARENT);
+            input.setLayoutParams(lp);
+            alertDialog.setView(input);
+            alertDialog.setPositiveButton("OK",
+                    new DialogInterface.OnClickListener() {
+                        public void onClick(DialogInterface dialog,int which) {
+                        	File file = new File(directory+selectedImg.getTitle());
+                        	file.renameTo(new File(directory+input.getText()+".jpeg"));
+                        	gridAdapter.clear();	                        	
+                        	gridAdapter = new GalleryGridViewAdapter(GalleryActivity.this, R.layout.row_grid, imageItems,directory,height,width);
+            	            gridView.setAdapter(gridAdapter);  
+                        	dialog.dismiss();
+                        }
+                    });
+            
+            alertDialog.setNegativeButton("Cancel",
+                    new DialogInterface.OnClickListener() {
+                        public void onClick(DialogInterface dialog, int which) {
+                            dialog.cancel();
+                        }
+                    });
+
+            alertDialog.show();
+
+        } else if (item == mItemSwitchDirectory) {
+        	if(directory.matches(".*myImages.*"))
+        		directory = Environment.getExternalStorageDirectory().getAbsolutePath()+"/soAppDir/myFilterImages/";
+        	else
+        		directory = Environment.getExternalStorageDirectory().getAbsolutePath()+"/soAppDir/myImages/";
+        	gridAdapter.clear();
+        	gridAdapter = new GalleryGridViewAdapter(this, R.layout.row_grid, imageItems,directory,height,width);
+            gridView.setAdapter(gridAdapter);   
+        	gridAdapter.notifyDataSetChanged();
+        } else if(item == mItemBTOptions) {
+        	startActivity(core.getPropertiesMenuEvents().runBTPropertiesActivity(this));
+        } else if(item == mItemFilterOptions) {
+        	startActivity(core.getPropertiesMenuEvents().runFilterPropertiesActivity(this));
+        } else if(item == mItemShowImg&&selectedImg!=null) {
+        	Intent intent = new Intent(Intent.ACTION_VIEW);
+        	File file = new File(directory+selectedImg.getTitle());
+        	Log.d("Gallery","show img: "+file.getAbsolutePath());
+        	intent.setDataAndType(Uri.fromFile(file),"image/*");
+        	startActivity(intent);
+        }    
+    	return true;
     }
     private BaseLoaderCallback mLoaderCallback = new BaseLoaderCallback(this) {
